worker_api/requirements.txt
pyyaml
docker==2.5.0
consulate
flask
flask-restful
